requirements.txt
streamlit
openai-whisper
yt-dlp
fpdf
ffmpeg-python